+++ src/ic0/src/ic0.rs
@@ -42,6 +42,7 @@ extern "C" {
     pub fn call_cycles_add(amount: i64);
     pub fn call_cycles_add128(amount_high: i64, amount_low: i64);
     pub fn call_perform() -> i32;
+    pub fn global_timer_set(timestamp: i64) -> i64;
     pub fn stable_size() -> i32;
     pub fn stable_grow(new_pages: i32) -> i32;
     pub fn stable_write(offset: i32, src: i32, size: i32);
@@ -163,6 +164,9 @@ mod non_wasm {
     }
     pub unsafe fn call_perform() -> i32 {
         panic!("call_perform should only be called inside canisters.");
+    }
+    pub unsafe fn global_timer_set(timestamp: i64) -> i64 {
+        panic!("global_timer_set should only be called inside canisters.");
     }
     pub unsafe fn stable_size() -> i32 {
         panic!("stable_size should only be called inside canisters.");